Added Following Pokemon mechanics and sprites/palettes for all Gen 1 Pokemon

+++ src/fldeff_rocksmash.c
@@ -29,7 +29,7 @@ static void FieldCallback_RockSmash(void);
 static void FieldMove_RockSmash(void);
 
 // text
-bool8 CheckObjectGraphicsInFrontOfPlayer(u8 graphicsId)
+bool8 CheckObjectGraphicsInFrontOfPlayer(u16 graphicsId)
 {
     u8 objEventId;
